LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/negative-overflow-003-expected.txt

1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3

PASS .container 1
PASS .container 2
FAIL .container 3 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: ltr; flex-flow: wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
FAIL .container 4 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-rl; direction: ltr; flex-flow: row-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 370 but got 360
FAIL .container 5 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: ltr; flex-flow: row-reverse wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 130 but got 120
FAIL .container 6 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: ltr; flex-flow: row-reverse wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
PASS .container 7
PASS .container 8
FAIL .container 9 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-rl; direction: ltr; flex-flow: column wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 370 but got 360
FAIL .container 10 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: ltr; flex-flow: column-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
FAIL .container 11 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-rl; direction: ltr; flex-flow: column-reverse wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 120
FAIL .container 12 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-rl; direction: ltr; flex-flow: column-reverse wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 120
FAIL .container 13 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-rl; direction: rtl; flex-flow: row;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 370 but got 360
FAIL .container 14 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: rtl; flex-flow: wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 130 but got 120
FAIL .container 15 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: rtl; flex-flow: wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
PASS .container 16
PASS .container 17
FAIL .container 18 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: rtl; flex-flow: row-reverse wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
FAIL .container 19 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: rtl; flex-flow: column;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 130 but got 120
FAIL .container 20 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-rl; direction: rtl; flex-flow: column wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 370 but got 360
PASS .container 21
FAIL .container 22 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-rl; direction: rtl; flex-flow: column-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
FAIL .container 23 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-rl; direction: rtl; flex-flow: column-reverse wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 120
FAIL .container 24 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-rl; direction: rtl; flex-flow: column-reverse wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 120
FAIL .container 25 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-lr; direction: ltr; flex-flow: row;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 370 but got 360
FAIL .container 26 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: ltr; flex-flow: wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 130 but got 120
FAIL .container 27 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: ltr; flex-flow: wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
PASS .container 28
PASS .container 29
FAIL .container 30 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: ltr; flex-flow: row-reverse wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
FAIL .container 31 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: ltr; flex-flow: column;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 130 but got 120
FAIL .container 32 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-lr; direction: ltr; flex-flow: column wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 370 but got 360
PASS .container 33
FAIL .container 34 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: ltr; flex-flow: column-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
FAIL .container 35 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-lr; direction: ltr; flex-flow: column-reverse wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 120
FAIL .container 36 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-lr; direction: ltr; flex-flow: column-reverse wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 120
PASS .container 37
PASS .container 38
FAIL .container 39 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: rtl; flex-flow: wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
FAIL .container 40 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-lr; direction: rtl; flex-flow: row-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 370 but got 360
FAIL .container 41 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: rtl; flex-flow: row-reverse wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 130 but got 120
FAIL .container 42 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: rtl; flex-flow: row-reverse wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
PASS .container 43
PASS .container 44
FAIL .container 45 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-lr; direction: rtl; flex-flow: column wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollHeight expected 370 but got 360
FAIL .container 46 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: sideways-lr; direction: rtl; flex-flow: column-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 360
FAIL .container 47 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-lr; direction: rtl; flex-flow: column-reverse wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 120
FAIL .container 48 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: sideways-lr; direction: rtl; flex-flow: column-reverse wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 120
